+++ frontend/src/components/RotateCredentials.vue
@@ -50,12 +50,11 @@ SPDX-License-Identifier: Apache-2.0
 import ActionButtonDialog from '@/components/dialogs/ActionButtonDialog'
 import { addShootAnnotation } from '@/utils/api'
 import { SnotifyPosition } from 'vue-snotify'
-import { shootStatusCredentialRotation, rotationTypes } from '@/mixins/shootStatusCredentialRotation'
+import shootStatusCredentialRotation from '@/mixins/shootStatusCredentialRotation'
 import { errorDetailsFromError } from '@/utils/error'
 import { mapGetters } from 'vuex'
 import includes from 'lodash/includes'
 import compact from 'lodash/compact'
-import map from 'lodash/map'
 
 export default {
   name: 'rotate-credentials',
@@ -65,13 +64,6 @@ export default {
   props: {
     text: {
       type: Boolean
-    },
-    type: {
-      type: String,
-      required: true,
-      validator: function (value) {
-        return map(rotationTypes, 'type').includes(value)
-      }
     }
   },
   data () {
